refactor: move selectors into logic
